update user controller and router
